+++ TRI/to_read.h
@@ -3,6 +3,6 @@
 #include <string>
 using namespace std;
 int charBitToInt(char a, int offset);
-void get_n(string aFile,int n, int* out);
+void get_n(string aFile,int n, float* out);
 int get_n_result(string aFile,int n);
 #endif //TO_READ_IMAGES
